CORRAIDZ-136 zfs: ztest needs to test dRAID zpools
This commit adds dRAID testing to ztest

This work was a collaboration between Sydney Vanda and Don Brady.

Included new command options to contol testing with dRAID:
 -K draid|raidz|random -- kind of RAID to test
 -D <value> -- dRAID data drives per redundancy group
 -G <value> -- dRAID redundancy group count
 -S <value> -- dRAID distributed spare drives
 -R <value> -- RAID parity

These values are used to generate a dRAID config from the draidcfg
command.  This also allows for more comprehensive configurations
when used with zloop.  By default, dRAID will be provisioned 50% of
the time, and 67% of the dRAID pools will use segregated metadata
and smallblks in a dRAID mirror.

Introduced a new option to control allocation class testing:
 -C dedicated|segregated=on|off|random

Fixed '-v vdevs' argument to actually limit total vdevs added to a pool.

Change-Id: I9e2349147d4183b31497b36edb9bb029156b5fde
Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>
Signed-off-by: Don Brady <don.brady@intel.com>
Reviewed-on: https://review.whamcloud.com/26108
Reviewed-by: John Salinas <John.Salinas@intel.com>
Tested-by: John Malmberg <john.e.malmberg@intel.com>